feat: Implement a modular LLM provider system with support for OpenAI, Gemini, Ollama, and browser-based LLMs, along with updated architecture documentation.
