+++ src/renderer/src/lib/llm.ts
@@ -1,15 +1,11 @@
 // LLM Orchestrator - Manages different LLM providers
 // Priority: WebLLM (On-Device) > Ollama > OpenAI-compatible
 
-import { FEATURE_FLAGS, LLM_CONFIG } from "./constants";
 import {
   getWebLLMStatus,
-  loadWebLLMModel,
-  chatWithWebLLM,
   subscribeToWebLLMStatus,
   WEBLLM_MODELS,
   type WebLLMStatus,
-  type WebLLMModelId,
   checkDownloadedWebLLMModels,
   deleteWebLLMModel,
   downloadWebLLMModelOnly,
@@ -17,18 +13,24 @@ import {
   checkWebLLMModelCompatibility
 } from "./webllm";
 import { CREATE_PLAN_TOOL } from "./plan_manager";
-import { EXECUTION_PLAN_SCHEMA } from "./agent-protocol";
-import { getUserEnvironmentContext } from "./user-environment";
 import {
   LLMMessage,
   LLMTool,
   ServerInfo,
   LLMResponse,
   LLMProvider,
-  LLMSettings,
-  LLMContentPart
+  LLMSettings
 } from "./types";
 import { pruneContext } from "./dcp";
+
+// Sub-modules
+import { ProviderStatus } from "./llm/types";
+import { checkOllama, testOllamaConnection, callOllama } from "./llm/ollama";
+import { checkBrowserLLM, testWebLLMConnection, callBrowserLLM, downloadBrowserModel } from "./llm/browser-llm";
+import { checkOpenAI, checkOpenRouter, testOpenAIConnection, callOpenAI } from "./llm/openai";
+import { checkGemini, testGeminiConnection, callGemini } from "./llm/gemini";
+import { buildSystemPrompt } from "./llm/prompts";
+import { ensureRecord, safeParseJSON } from "./llm/utils";
 
 export {
   getWebLLMStatus,
@@ -46,525 +48,26 @@ export {
   type ServerInfo,
   type LLMResponse,
   type LLMProvider,
-  type LLMSettings
+  type LLMSettings,
+  // Provider health checks
+  checkOllama,
+  testOllamaConnection,
+  checkBrowserLLM,
+  testWebLLMConnection,
+  checkOpenAI,
+  checkOpenRouter,
+  checkGemini,
+  testGeminiConnection,
+  testOpenAIConnection,
+  downloadBrowserModel,
+  // Shared utilities
+  ensureRecord,
+  safeParseJSON
 };
 
-interface ProviderStatus {
-
-  available: boolean;
-  model?: string;
-  error?: string;
-  models?: string[]; // Available models list
-  modelsEndpointAvailable?: boolean; // Whether /models endpoint exists
-  // WebLLM specific
-  isWebGPUSupported?: boolean;
-  isLoaded?: boolean;
-  isLoading?: boolean;
-  loadingProgress?: number;
-  loadingStage?: string;
-  downloadedModels?: string[]; // WebLLM cached models
-}
-
-// Get Ollama settings from store or use defaults
-function getOllamaSettings(settings?: LLMSettings) {
-  const baseUrl = settings?.ollamaBaseUrl || LLM_CONFIG.OLLAMA.BASE_URL;
-  const model = settings?.ollamaModel || LLM_CONFIG.OLLAMA.DEFAULT_MODEL;
-  return { baseUrl, model };
-}
-
-// Get OpenAI settings from store or use defaults
-async function getOpenAISettings(
-  settings?: LLMSettings
-): Promise<{ apiKey: string; baseUrl: string; model: string }> {
-  // Import electron here to avoid circular dependencies
-  const electron = (await import("./electron")).default;
-
-  const apiKey =
-    settings?.openaiApiKey ||
-    (await electron.secure.get("openai_api_key")).value ||
-    "";
-  const baseUrl =
-    settings?.openaiBaseUrl ||
-    (await electron.store.get<string>("openai_base_url")) ||
-    "https://api.openai.com/v1";
-  const model =
-    settings?.openaiModel || LLM_CONFIG.OPENAI_COMPATIBLE.DEFAULT_MODEL;
-  return { apiKey, baseUrl, model };
-}
-
-// Get Gemini settings from store or use defaults
-async function getGeminiSettings(
-  settings?: LLMSettings
-): Promise<{ apiKey: string; baseUrl: string; model: string }> {
-  const electron = (await import("./electron")).default;
-  const apiKey =
-    settings?.geminiApiKey ||
-    (await electron.secure.get("gemini_api_key")).value ||
-    "";
-  const baseUrl = LLM_CONFIG.GEMINI.BASE_URL;
-  const model = settings?.geminiModel || LLM_CONFIG.GEMINI.DEFAULT_MODEL;
-  return { apiKey, baseUrl, model };
-}
-
-// Get OpenRouter settings from store or use defaults
-async function getOpenRouterSettings(
-  settings?: LLMSettings
-): Promise<{ apiKey: string; baseUrl: string; model: string }> {
-  const electron = (await import("./electron")).default;
-  const apiKey =
-    settings?.openrouterApiKey ||
-    (await electron.secure.get("openrouter_api_key")).value ||
-    "";
-  const baseUrl = LLM_CONFIG.OPENROUTER.BASE_URL;
-  const model = settings?.openrouterModel || LLM_CONFIG.OPENROUTER.DEFAULT_MODEL;
-  return { apiKey, baseUrl, model };
-}
-
-// Check if Ollama is running and list available models
-export async function checkOllama(
-  settings?: LLMSettings
-): Promise<ProviderStatus> {
-  if (!FEATURE_FLAGS.OLLAMA_ENABLED) {
-    return { available: false, error: "Ollama disabled" };
-  }
-
-  const { baseUrl, model: preferredModel } = getOllamaSettings(settings);
-
-  try {
-    // Use IPC to fetch models from main process (bypasses potential CORS issues)
-    const electron = (window as any).electron;
-    if (electron?.llm?.fetchOllamaModels) {
-      const result = await electron.llm.fetchOllamaModels(baseUrl);
-
-      if (result.success && result.models && result.models.length > 0) {
-        const defaultModel =
-          result.models.find((m: string) => m.startsWith(preferredModel)) ||
-          result.models[0] ||
-          preferredModel;
-        return {
-          available: true,
-          model: defaultModel,
-          models: result.models,
-        };
-      } else if (result.error) {
-        return { available: false, error: result.error };
-      }
-      return { available: false, error: "No models found" };
-    }
-
-    // Fallback to direct fetch if IPC not available (for development/testing)
-    const response = await fetch(`${baseUrl}/api/tags`);
-    if (response.ok) {
-      const data = await response.json();
-      const models = (data.models || []).map((m: { name: string }) => m.name);
-      const defaultModel =
-        models.find((m: string) => m.startsWith(preferredModel)) ||
-        models[0] ||
-        preferredModel;
-      return {
-        available: true,
-        model: defaultModel,
-        models: models,
-      };
-    }
-    return { available: false, error: "Ollama not responding" };
-  } catch (error) {
-    return {
-      available: false,
-      error: error instanceof Error ? error.message : "Ollama not running",
-    };
-  }
-}
-
-// Test Ollama connection with a specific model
-export async function testOllamaConnection(
-  baseUrl: string,
-  model: string
-): Promise<{ success: boolean; error?: string }> {
-  try {
-    const response = await fetch(`${baseUrl}/api/generate`, {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        model,
-        prompt: "test",
-        stream: false,
-      }),
-    });
-    if (response.ok) {
-      return { success: true };
-    }
-    const error = await response.json().catch(() => ({}));
-    return { success: false, error: error.error || "Connection failed" };
-  } catch (error) {
-    return {
-      success: false,
-      error: error instanceof Error ? error.message : "Connection failed",
-    };
-  }
-}
-
-// Check if WebLLM (On-Device AI via WebGPU) is available
-export async function checkBrowserLLM(): Promise<ProviderStatus> {
-  if (!FEATURE_FLAGS.BROWSER_LLM_ENABLED) {
-    return { available: false, error: 'On-Device AI disabled' }
-  }
-
-  try {
-    const status = getWebLLMStatus();
-    console.log('[WebLLM] Status:', status);
-
-    if (!status.isSupported) {
-      return {
-        available: false,
-        error: status.error || 'WebGPU not supported',
-        isWebGPUSupported: false,
-      }
-    }
-
-    // WebGPU is supported
-    const models = WEBLLM_MODELS.map(m => m.id);
-
-    return {
-      available: true,
-      model: status.currentModel || WEBLLM_MODELS[0].id,
-      models: models,
-      isWebGPUSupported: true,
-      isLoaded: status.isLoaded,
-      isLoading: status.isLoading,
-      loadingProgress: status.loadingProgress,
-      loadingStage: status.loadingStage,
-      downloadedModels: status.downloadedModels,
-      error: status.error || undefined,
-    }
-  } catch (error) {
-    console.error('[WebLLM] Check error:', error);
-    return {
-      available: false,
-      error: error instanceof Error ? error.message : 'WebLLM check failed'
-    }
-  }
-}
-
-// Test WebLLM connection (simple chat)
-export async function testWebLLMConnection(): Promise<{ success: boolean; error?: string }> {
-  try {
-    const status = getWebLLMStatus();
-    if (!status.isLoaded) {
-      return { success: false, error: 'Model not loaded' };
-    }
-
-    // specific test message to avoid long responses
-    const response = await chatWithWebLLM([
-      { role: 'user', content: 'Say "test" and nothing else.' }
-    ]);
-
-    if (response.content) {
-      return { success: true };
-    }
-    return { success: false, error: 'No response content' };
-  } catch (error) {
-    return {
-      success: false,
-      error: error instanceof Error ? error.message : 'Test failed'
-    };
-  }
-}
-
-// Check if OpenAI-compatible API is configured and fetch available models
-export async function checkOpenAI(
-  settings?: LLMSettings,
-  providerOverride?: "openai" | "openrouter"
-): Promise<ProviderStatus> {
-  if (!FEATURE_FLAGS.CLOUD_LLM_ENABLED) {
-    return { available: false, error: "Cloud LLM disabled" };
-  }
-
-  const { apiKey, baseUrl, model } =
-    providerOverride === "openrouter"
-      ? await getOpenRouterSettings(settings)
-      : await getOpenAISettings(settings);
-
-  if (!apiKey) {
-    return {
-      available: false,
-      error: `${providerOverride === "openrouter" ? "OpenRouter" : "OpenAI"} API Key not set`,
-    };
-  }
-
-  try {
-    // Use IPC to fetch models from main process (bypasses CORS)
-    const electron = (window as any).electron;
-    if (electron?.llm?.fetchOpenAIModels) {
-      const result = await electron.llm.fetchOpenAIModels(baseUrl, apiKey);
-
-      if (result.success && result.models && result.models.length > 0) {
-        const models = result.models;
-        // Find the preferred model or use default
-        const preferredModel = model;
-        const defaultModel =
-          models.find((m: string) => m === preferredModel) ||
-          models.find((m: string) => m.includes("gpt-4o")) ||
-          models.find((m: string) => m.includes("gpt-4")) ||
-          models[0] ||
-          preferredModel;
-
-        return {
-          available: true,
-          model: defaultModel,
-          models: models,
-          modelsEndpointAvailable: true,
-        };
-      } else {
-        // If models endpoint fails, still mark as available if API key exists
-        return {
-          available: true,
-          model: model,
-          models: [model], // Fallback to just the configured model
-          modelsEndpointAvailable: false,
-          error: result.error || "Could not fetch models list",
-        };
-      }
-    } else {
-      // Fallback to direct fetch if IPC not available (for development/testing)
-      console.log(`[LLM Check] Checking OpenAI models at: ${baseUrl}/models`);
-      const response = await fetch(`${baseUrl}/models`, {
-        method: "GET",
-        headers: {
-          Authorization: `Bearer ${apiKey}`,
-        },
-      });
-
-      if (!response) {
-        throw new Error("Fetch returned undefined response");
-      }
-
-      if (response.ok) {
-        const data = await response.json();
-        const models = (data.data || [])
-          .filter((m: { id: string }) => {
-            const id = m.id.toLowerCase();
-            return (
-              id.includes("gpt") ||
-              id.includes("chat") ||
-              id.includes("claude") ||
-              id.includes("llama") ||
-              id.includes("perplexity")
-            );
-          })
-          .map((m: { id: string }) => m.id)
-          .sort();
-
-        const preferredModel = model;
-        const defaultModel =
-          models.find((m: string) => m === preferredModel) ||
-          models.find((m: string) => m.includes("gpt-4o")) ||
-          models.find((m: string) => m.includes("gpt-4")) ||
-          models[0] ||
-          preferredModel;
-
-        return {
-          available: true,
-          model: defaultModel,
-          models: models,
-          modelsEndpointAvailable: true,
-        };
-      } else {
-        if (response.status === 404) {
-          console.error(`[LLM Check] 404 Error: Models endpoint not found at ${baseUrl}/models`);
-        } else {
-          console.error(`[LLM Check] HTTP Error ${response.status}: ${response.statusText} at ${baseUrl}/models`);
-        }
-        return {
-          available: true,
-          model: model,
-          models: [model],
-          modelsEndpointAvailable: false,
-        };
-      }
-    }
-  } catch (error) {
-    // If fetch fails, still mark as available if API key exists
-    // User can still use the manually entered model
-    return {
-      available: true,
-      model: model,
-      models: [model], // Fallback to just the configured model
-      modelsEndpointAvailable: false,
-      error:
-        error instanceof Error ? error.message : "Could not fetch models list",
-    };
-  }
-}
-
-// Check if OpenRouter is configured and available
-export async function checkOpenRouter(
-  settings?: LLMSettings
-): Promise<ProviderStatus> {
-  const { apiKey, baseUrl, model } = await getOpenRouterSettings(settings);
-  if (!apiKey) return { available: false, error: "OpenRouter API Key not set" };
-
-  // reuse OpenAI check with OpenRouter specific headers if needed
-  return checkOpenAI(settings, "openrouter");
-}
-
-// Check if Gemini is configured and available
-export async function checkGemini(
-  settings?: LLMSettings
-): Promise<ProviderStatus> {
-  const { apiKey, model } = await getGeminiSettings(settings);
-  if (!apiKey) return { available: false, error: "Gemini API Key not set" };
-
-  try {
-    const baseUrl = LLM_CONFIG.GEMINI.BASE_URL;
-    const response = await fetch(`${baseUrl}/models?key=${apiKey}`);
-
-    if (response.ok) {
-      const data = await response.json();
-      const models = (data.models || [])
-        .filter((m: { name: string }) => m.name.includes("gemini"))
-        .map((m: { name: string }) => m.name.split("/").pop())
-        .filter(Boolean) as string[];
-
-      return {
-        available: true,
-        model: models.find(m => m === model) || models[0] || model,
-        models: models,
-        modelsEndpointAvailable: true,
-      };
-    }
-    return {
-      available: true,
-      model: model,
-      models: [model],
-      modelsEndpointAvailable: false,
-      error: "Could not fetch Gemini models list",
-    };
-  } catch (error) {
-    return {
-      available: true,
-      model: model,
-      models: [model],
-      modelsEndpointAvailable: false,
-    };
-  }
-}
-
-// Test Gemini connection and fetch models
-export async function testGeminiConnection(
-  apiKey: string,
-  model: string
-): Promise<{
-  success: boolean;
-  error?: string;
-  models?: string[];
-  modelsEndpointAvailable?: boolean;
-}> {
-  try {
-    const baseUrl = LLM_CONFIG.GEMINI.BASE_URL;
-    const response = await fetch(`${baseUrl}/models?key=${apiKey}`);
-
-    if (response.ok) {
-      const data = await response.json();
-      const models = (data.models || [])
-        .filter((m: { name: string }) => m.name.includes("gemini"))
-        .map((m: { name: string }) => m.name.split("/").pop())
-        .filter(Boolean) as string[];
-
-      return {
-        success: true,
-        models,
-        modelsEndpointAvailable: true,
-      };
-    } else {
-      const error = await response.json().catch(() => ({}));
-      return {
-        success: false,
-        error: error.error?.message || "Connection failed",
-      };
-    }
-  } catch (error) {
-    return {
-      success: false,
-      error: error instanceof Error ? error.message : "Connection failed",
-    };
-  }
-}
-
-// Test OpenAI connection and fetch models
-export async function testOpenAIConnection(
-  baseUrl: string,
-  apiKey: string,
-  model: string
-): Promise<{
-  success: boolean;
-  error?: string;
-  models?: string[];
-  modelsEndpointAvailable?: boolean;
-}> {
-  try {
-    // Test connection with chat completions
-    const response = await fetch(`${baseUrl}/chat/completions`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${apiKey}`,
-      },
-      body: JSON.stringify({
-        model,
-        messages: [{ role: "user", content: "test" }],
-        max_tokens: 5,
-      }),
-    });
-
-    if (!response.ok) {
-      const error = await response.json().catch(() => ({}));
-      return {
-        success: false,
-        error: error.error?.message || "Connection failed",
-      };
-    }
-
-    // Try to fetch models if connection succeeds
-    const electron = (window as any).electron;
-    if (electron?.llm?.fetchOpenAIModels) {
-      try {
-        const modelsResult = await electron.llm.fetchOpenAIModels(
-          baseUrl,
-          apiKey
-        );
-        if (modelsResult.success && modelsResult.models) {
-          return {
-            success: true,
-            models: modelsResult.models,
-            modelsEndpointAvailable: true,
-          };
-        } else {
-          return {
-            success: true,
-            modelsEndpointAvailable: false,
-            error: modelsResult.error || "Models endpoint not available",
-          };
-        }
-      } catch (modelsError) {
-        // Connection works but models endpoint doesn't
-        return {
-          success: true,
-          modelsEndpointAvailable: false,
-        };
-      }
-    }
-
-    return { success: true, modelsEndpointAvailable: false };
-  } catch (error) {
-    return {
-      success: false,
-      error: error instanceof Error ? error.message : "Connection failed",
-    };
-  }
-}
-
-// Get available providers
+/**
+ * Get available providers and their status
+ */
 export async function getAvailableProviders(
   settings?: LLMSettings
 ): Promise<Record<LLMProvider, ProviderStatus>> {
@@ -591,733 +94,9 @@ export async function getAvailableProviders(
   };
 }
 
-// Helper to extract text from multimodal content for providers that only support text
-function extractTextForLegacyProviders(content: string | LLMContentPart[]): string {
-  if (typeof content === 'string') return content;
-  return content.map(p => p.type === 'text' ? p.text : '').join('\n');
-}
-
-// Call Ollama API
-async function callOllama(
-  messages: LLMMessage[],
-  tools?: LLMTool[],
-  settings?: LLMSettings,
-  abortSignal?: AbortSignal,
-  workspacePath?: string // New parameter
-): Promise<LLMResponse> {
-  const { baseUrl, model } = getOllamaSettings(settings);
-
-  console.log(`[LLM Chat] Calling Ollama at: ${baseUrl}/api/chat`);
-  const response = await fetch(`${baseUrl}/api/chat`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    signal: abortSignal,
-    body: JSON.stringify({
-      model: model,
-      messages: messages.map((m) => ({
-        role: m.role,
-        content: extractTextForLegacyProviders(m.content),
-      })),
-      stream: false,
-      tools: tools?.map((t) => ({
-        type: "function",
-        function: {
-          name: t.name,
-          description: t.description,
-          parameters: t.parameters,
-        },
-      })),
-    }),
-  });
-
-  if (!response || !response.ok) {
-    if (response?.status === 404) {
-      console.error(`[LLM Chat] 404 Error: Ollama endpoint not found at ${baseUrl}/api/chat`);
-    }
-    const error = response ? await response.json().catch(() => ({})) : { error: "Fetch failed" };
-    throw new Error(error.error || `Ollama error: ${response?.statusText || "Unknown"}`);
-  }
-
-  const data = await response.json();
-
-  return {
-    content: data.message?.content || "",
-    toolCalls: data.message?.tool_calls?.map(
-      (tc: {
-        function: { name: string; arguments: Record<string, unknown> };
-      }) => ({
-        id: `call_${Date.now()}`,
-        name: tc.function.name,
-        arguments: tc.function.arguments,
-      })
-    ),
-    provider: "ollama",
-    model: model,
-  };
-}
-
-// Call WebLLM (On-Device AI via WebGPU)
-// Note: Most WebLLM models don't support native tool calling, so we rely on
-// the system prompt to instruct the model to output JSON, then parse it
-async function callBrowserLLM(
-  messages: LLMMessage[],
-  tools?: LLMTool[],
-  settings?: LLMSettings,
-  workspacePath?: string // New parameter 
-): Promise<LLMResponse> {
-  try {
-    const status = getWebLLMStatus();
-
-    // Check if model is loaded
-    if (!status.isLoaded) {
-      console.log('[WebLLM] Model not loaded, attempting to load...');
-      await loadWebLLMModel();
-    }
-
-    // Don't pass tools to WebLLM - the system prompt already contains tool definitions
-    // Models will output JSON tool calls in their response content
-    const response = await chatWithWebLLM(
-      messages
-        .filter(m => m.role !== 'tool')
-        .map(m => ({
-          role: m.role as 'user' | 'assistant' | 'system',
-          content: extractTextForLegacyProviders(m.content)
-        }))
-      // No tools passed - avoids "model doesn't support tools" error
-    );
-
-    // Parse tool calls from JSON in response content (same as existing JSON fallback)
-    let toolCalls = response.toolCalls;
-    if (!toolCalls && response.content && tools && tools.length > 0) {
-      toolCalls = parseToolCallsFromJson(response.content);
-    }
-
-    return {
-      content: response.content,
-      toolCalls,
-      provider: 'browser',
-      model: status.currentModel || 'webllm',
-    };
-  } catch (error) {
-    console.error('[WebLLM] Chat error:', error);
-    throw new Error(`WebLLM error: ${error instanceof Error ? error.message : 'Unknown error'}`);
-  }
-}
-
-
-// Helper to format messages for OpenAI-compatible APIs
-function formatMessagesForOpenAI(messages: LLMMessage[]): any[] {
-  return messages.map(m => {
-    // Basic message structure
-    const formatted: any = {
-      role: m.role,
-      content: m.content
-    };
-
-    // Add tool_calls if present (and strictly stringify arguments)
-    if (m.tool_calls && m.tool_calls.length > 0) {
-      formatted.tool_calls = m.tool_calls.map((tc: any) => ({
-        id: tc.id,
-        type: 'function',
-        function: {
-          name: tc.function.name,
-          arguments: typeof tc.function.arguments === 'object'
-            ? JSON.stringify(tc.function.arguments)
-            : tc.function.arguments
-        }
-      }));
-    }
-
-    // Add tool_call_id if it's a tool response
-    if (m.role === 'tool' && m.tool_call_id) {
-      formatted.tool_call_id = m.tool_call_id;
-    }
-
-    // Add name if present (formatting for OpenAI)
-    if (m.name) {
-      formatted.name = m.name;
-    }
-
-    return formatted;
-  });
-}
-
-// Call OpenAI-compatible API
-async function callOpenAI(
-  messages: LLMMessage[],
-  tools?: LLMTool[],
-  settings?: LLMSettings,
-  useJsonFallback: boolean = false,
-  servers?: ServerInfo[],
-  isOpenRouter: boolean = false,
-  abortSignal?: AbortSignal,
-  dynamicRules?: string,
-  isSubAgent?: boolean,
-  workspacePath?: string // New parameter
-): Promise<LLMResponse> {
-  const { apiKey, baseUrl, model } = isOpenRouter
-    ? await getOpenRouterSettings(settings)
-    : await getOpenAISettings(settings);
-
-  if (!apiKey) {
-    throw new Error("OpenAI API key not configured");
-  }
-
-  const headers: Record<string, string> = {
-    "Content-Type": "application/json",
-    Authorization: `Bearer ${apiKey}`,
-  };
-
-  if (isOpenRouter) {
-    headers["HTTP-Referer"] = "https://ai-worker.app";
-    headers["X-Title"] = "AI-Worker";
-  }
-
-  // If using JSON fallback, rebuild system message with JSON instructions
-  let requestMessages = messages;
-  if (useJsonFallback && tools && tools.length > 0) {
-    requestMessages = [...messages];
-    // Rebuild system message with JSON fallback instructions
-    const systemMsgIndex = requestMessages.findIndex(
-      (m) => m.role === "system"
-    );
-    const systemPrompt = await buildSystemPrompt(tools, servers, true);
-    if (systemMsgIndex >= 0) {
-      requestMessages[systemMsgIndex] = {
-        role: "system" as const,
-        content: systemPrompt,
-      };
-    } else {
-      requestMessages.unshift({
-        role: "system" as const,
-        content: systemPrompt,
-      });
-    }
-  }
-
-  console.log(`[LLM Chat] Calling OpenAI-compatible API at: ${baseUrl}/chat/completions`);
-  const response = await fetch(`${baseUrl}/chat/completions`, {
-    method: "POST",
-    headers,
-    signal: abortSignal,
-    body: JSON.stringify({
-      model: model,
-      messages: formatMessagesForOpenAI(useJsonFallback ? requestMessages : messages),
-      ...(useJsonFallback
-        ? {}
-        : {
-          tools: tools?.map((t) => ({
-            type: "function",
-            function: {
-              name: t.name,
-              description: t.description,
-              parameters: t.parameters,
-            },
-          })),
-        }),
-    }),
-  });
-
-  if (!response || !response.ok) {
-    if (response?.status === 404) {
-      console.error(`[LLM Chat] 404 Error: OpenAI-compatible endpoint not found at ${baseUrl}/chat/completions`);
-    }
-    const error = response ? await response.json().catch(() => ({})) : { error: "Fetch failed" };
-    const errorMessage =
-      error.error?.message || `OpenAI error: ${response?.statusText || "Unknown"}`;
-
-    // Check if it's a tool calling not supported error
-    if (
-      error.error?.code === 400 &&
-      (errorMessage.toLowerCase().includes("tool calling") ||
-        errorMessage.toLowerCase().includes("tool_call") ||
-        errorMessage.toLowerCase().includes("tools")) &&
-      !useJsonFallback &&
-      tools &&
-      tools.length > 0
-    ) {
-      // Retry with JSON fallback - rebuild messages with proper system prompt
-      const retryMessages = [...messages];
-      const systemMsgIndex = retryMessages.findIndex(
-        (m) => m.role === "system"
-      );
-      const systemPrompt = await buildSystemPrompt(tools, servers, true);
-      if (systemMsgIndex >= 0) {
-        retryMessages[systemMsgIndex] = {
-          role: "system" as const,
-          content: systemPrompt,
-        };
-      } else {
-        retryMessages.unshift({
-          role: "system" as const,
-          content: systemPrompt,
-        });
-      }
-      return callOpenAI(retryMessages, tools, settings, true, servers);
-    }
-
-    throw new Error(errorMessage);
-  }
-
-  const data = await response.json();
-
-  if (!data.choices || !Array.isArray(data.choices) || data.choices.length === 0) {
-    if (data.error) {
-      throw new Error(data.error.message || JSON.stringify(data.error));
-    }
-    throw new Error(`Unexpected API response format: ${JSON.stringify(data)}`);
-  }
-
-  const choice = data.choices[0];
-  const content = choice.message?.content || "";
-
-  // If using JSON fallback, try to parse tool calls from content
-  let toolCalls = choice.message?.tool_calls?.map(
-    (tc: { id: string; function: { name: string; arguments: string } }) => {
-      let args: any = {};
-
-      // Handle null, undefined, or empty string arguments
-      if (!tc.function.arguments || tc.function.arguments === 'null' || tc.function.arguments === 'undefined') {
-        console.warn(`[LLM] Tool call "${tc.function.name}" has null/empty arguments. Using empty object.`);
-        args = {};
-      } else {
-        try {
-          const parsed = JSON.parse(tc.function.arguments);
-          // JSON.parse(null) returns null, so we need to check the result
-          args = parsed !== null && typeof parsed === 'object' ? parsed : {};
-        } catch (e) {
-          console.warn(`[LLM] Failed to parse tool arguments for "${tc.function.name}":`, tc.function.arguments);
-          args = { _parse_error: "Invalid JSON arguments from LLM" };
-        }
-      }
-
-      console.log(`[LLM] Native Tool Call Identified: ${tc.function.name}`, {
-        tool: tc.function.name,
-        arguments: args,
-        raw: tc.function.arguments
-      });
-
-      return {
-        id: tc.id,
-        name: tc.function.name,
-        arguments: ensureRecord(args), // Use parsed args
-      };
-    }
-  );
-
-  // If no native tool calls, try to parse from content (Self-Healing)
-  if (!toolCalls || toolCalls.length === 0) {
-    if (useJsonFallback && content) {
-      console.log('[LLM] No native tool calls found. Attempting to parse JSON from content...');
-      toolCalls = parseToolCallsFromJson(content);
-      if (toolCalls && toolCalls.length > 0) {
-        console.log(`[LLM] Successfully recovered ${toolCalls.length} tool calls from content body.`);
-      }
-    }
-
-    // Check for XML Plan (Legacy/Model Hallucination Fallback)
-    else if (content.includes('<agent_plan>')) {
-      // ... existing XML logic ...
-      console.log('[LLM] Detected XML plan in content, converting to tool call');
-      toolCalls = [{
-        id: `auto_plan_${Date.now()}`,
-        name: 'create_execution_plan',
-        arguments: { plan_content: content }
-      }];
-    }
-  }
-
-  return {
-    content: content,
-    toolCalls: toolCalls,
-    provider: "openai",
-    model: model,
-  };
-}
-
-// Parse tool calls from JSON in response content
-function parseToolCallsFromJson(
-  content: string
-): LLMResponse["toolCalls"] | undefined {
-  try {
-    // Try to find JSON in the content (might be in code blocks or raw)
-    let jsonStr = content.trim();
-
-    // Remove markdown code blocks if present
-    jsonStr = jsonStr
-      .replace(/```json\n?/g, "")
-      .replace(/```\n?/g, "")
-      .trim();
-
-    // Try to extract JSON object
-    const jsonMatch = jsonStr.match(/\{[\s\S]*\}/);
-    if (jsonMatch) {
-      const parsed = JSON.parse(jsonMatch[0]);
-
-      // Standard Format: { "tool_calls": [...] }
-      if (parsed.tool_calls && Array.isArray(parsed.tool_calls)) {
-        console.log('[LLM] Identified Standard JSON Tool Call Format');
-        return parsed.tool_calls.map(
-          (
-            tc: { name: string; arguments: Record<string, unknown> },
-            idx: number
-          ) => {
-            console.log(`[LLM] JSON Recovered Call: ${tc.name}`, tc.arguments);
-            return {
-              id: `json_call_${Date.now()}_${idx}`,
-              name: tc.name,
-              arguments: ensureRecord(tc.arguments),
-            };
-          }
-        );
-      }
-
-      // Alternate Format (Common in some models): { "tool": "name", "params": {...} }
-      if (parsed.tool && typeof parsed.tool === 'string') {
-        console.log('[LLM] Identified Alternate JSON Tool Call Format:', parsed.tool);
-        // Normalize parameters: some models use params, parameters, arguments, or even 'args'
-        const params = parsed.params || parsed.parameters || parsed.arguments || parsed.args || {};
-        console.log(`[LLM] Normalized parameters for "${parsed.tool}":`, params);
-
-        return [{
-          id: `json_call_${Date.now()}`,
-          name: parsed.tool,
-          arguments: ensureRecord(params)
-        }];
-      }
-    }
-  } catch (error) {
-    // Failed to parse, return undefined
-    console.warn("Failed to parse tool calls from JSON:", error);
-  }
-  return undefined;
-}
-
 /**
- * Filter tools to most relevant subset for sub-agents (reduces token usage)
- * Prioritizes common automation tools across browser, file, API, and messaging categories
+ * Main chat function - automatically selects best provider based on configuration
  */
-//TODO: MESSAGE FOR LLM/CODING AGENT : this function need to be uodated based on the mcp servers connected by default in code 
-function filterRelevantTools(tools?: LLMTool[], taskHint?: string): LLMTool[] {
-  if (!tools || tools.length <= 20) return tools || [];
-
-  // Priority patterns for common automation tasks (tool-agnostic)
-  const priorityPatterns = [
-    // Browser/UI operations (Playwright MCP)
-    /browser|navigate|click|type|screenshot|snapshot|goto|page|playwright/i,
-    /fill|submit|select|input|press|scroll|wait|hover/i,
-    // File operations
-    /file|read|write|create|delete|copy|move|list|directory|folder/i,
-    // API/HTTP operations
-    /api|http|request|fetch|post|get|put|patch|endpoint|webhook/i,
-    // Database operations
-    /database|db|query|sql|insert|update|select|table/i,
-    // Messaging/Communication
-    /message|send|email|slack|notification|chat|discord/i,
-    // Data extraction/manipulation
-    /search|find|get|extract|parse|convert|transform/i,
-    // State/Context
-    /state|status|info|current|context/i,
-  ];
-
-  // Score and sort tools by relevance
-  const scored = tools.map(t => {
-    let score = 0;
-    const nameAndDesc = `${t.name} ${t.description || ''}`.toLowerCase();
-
-    for (const pattern of priorityPatterns) {
-      if (pattern.test(nameAndDesc)) {
-        score += 10;
-      }
-    }
-
-    // Boost if task hint matches tool description
-    if (taskHint && nameAndDesc.includes(taskHint.toLowerCase().substring(0, 20))) {
-      score += 5;
-    }
-
-    return { tool: t, score };
-  });
-
-  // Return top 20 most relevant tools (increased from 15 for broader workflows)
-  return scored
-    .sort((a, b) => b.score - a.score)
-    .slice(0, 20)
-    .map(s => s.tool);
-}
-
-/**
- * Build a compact system prompt for sub-agents (~70% smaller than main prompt)
- * Includes essential rules (think tags, autonomous behavior) but removes verbose examples
- */
-async function buildSubAgentSystemPrompt(
-  tools?: LLMTool[],
-  dynamicRules?: string,
-  workspacePath?: string
-): Promise<string> {
-  // Filter to most relevant tools
-  const relevantTools = filterRelevantTools(tools, dynamicRules);
-  const toolCount = relevantTools?.length || 0;
-
-  // Compact tool list with descriptions (max 60 chars each)
-  const toolList = relevantTools?.map(t => {
-    const desc = (t.description || '').substring(0, 60);
-    return `- **${t.name}**: ${desc}${(t.description || '').length > 60 ? '...' : ''}`;
-  }).join('\n') || 'No tools';
-
-  const userContext = await getUserEnvironmentContext();
-
-  return `You are a focused sub-agent executing a delegated task.
-
-${userContext}
-
-${workspacePath ? `ACTIVE WORKSPACE: ${workspacePath}
-All filesystem operations (fs_*) MUST be performed within this directory.` : `WORKSPACE NOT SELECTED: 
-If the user's request involves filesystem operations (fs_*), explain that no workspace is selected and they should use the folder icon in the UI to select one.`}
-
-AVAILABLE TOOLS (${toolCount}):
-${toolList}
-
-# RESPONSE FORMAT
-Use <think>...</think> for reasoning (hidden). Put actions and final response outside <think> tags.
-
-# CORE RULES
-1. **Act, don't explain**: Use tools immediately, don't describe your plan
-2. **Be autonomous**: Don't ask permission, make decisions
-3. **Be concise**: Max 100 words in final response
-4. **Complete the current step**: Focus on what's asked, don't do extra steps
-5. **Error handling**: If tool fails, try alternative once, then report
-6. **Panic Mode**: If you are stuck for 3 turns, run get_state (or snapshot) and report findings.
-7. **End marker**: Finish with "✓ Done"
-${dynamicRules ? `\n# TASK-SPECIFIC\n${dynamicRules}` : ''}`;
-}
-
-// Build robust but token-efficient system prompt
-async function buildSystemPrompt(
-  tools?: LLMTool[],
-  servers?: ServerInfo[],
-  useJsonFallback = false,
-  dynamicRules?: string,
-  isSubAgent = false, // NEW: Flag for lightweight prompt
-  workspacePath?: string // Injected workspace path for filesystem scoping
-): Promise<string> {
-  // Use compact prompt for sub-agents
-  if (isSubAgent) {
-    return buildSubAgentSystemPrompt(tools, dynamicRules, workspacePath);
-  }
-
-  const toolCount = tools?.length || 0;
-  const serverCount = servers?.length || 0;
-
-  if (toolCount === 0) {
-    return `You are AI - Worker, a helpful voice - first assistant.When tools become available, use them to perform actions instead of providing manual instructions.Be concise for voice output.`;
-  }
-
-  // Ensure we have tools - this should never happen if tools are passed correctly
-  if (!tools || tools.length === 0) {
-    console.warn("buildSystemPrompt called with empty tools array");
-    return `You are AI - Worker, a helpful voice - first assistant.When tools become available, use them to perform actions instead of providing manual instructions.Be concise for voice output.`;
-  }
-
-  // Add JSON format instruction if using fallback
-  const jsonFormatNote = useJsonFallback
-    ? `\n\n ** CRITICAL: JSON TOOL CALLING FORMAT **\nThis model doesn't support native tool calling. When you need to use a tool, return ONLY a JSON object (no markdown, no code blocks, just raw JSON).\n\nFor the 'create_execution_plan' tool, use this exact structure:\n${JSON.stringify(EXECUTION_PLAN_SCHEMA, null, 2)}\n\nIMPORTANT: \n- Return ONLY the JSON object\n- Do not include any text before or after`
-    : "";
-
-  // Build tools description - compact format with name and description
-  const toolsDescription =
-    tools
-      ?.map((tool, idx) => {
-        // Extract key parameters for context (if available)
-        const params = tool.parameters as
-          | { properties?: Record<string, unknown>; required?: string[] }
-          | undefined;
-        const properties = params?.properties || {};
-        const paramNames = Object.keys(properties).slice(0, 3).join(", ");
-        const paramHint = paramNames
-          ? ` (params: ${paramNames}${Object.keys(properties).length > 3 ? "..." : ""
-          })`
-          : "";
-
-        const server = servers?.find(s => s.toolCount > 0 && tools?.some(t => t.name.startsWith(tool.name.split('_')[0])));
-        // Heuristic: check if we can query mcp.ts directly or pass server mapping. 
-        // Since we don't have direct mapping here, we can rely on grouping by server context below or just hint.
-        // Better: The 'servers' list passed to this function usually contains aggregate info. 
-        // Let's simplified: The "Connected MCP Servers" section below handles the grouping.
-        // We will just leave the tool description as is, but emphasize the Agent Roles above.
-
-        return `${idx + 1}. **${tool.name}**${paramHint}: ${tool.description}`;
-      })
-      .join("\n\n") || "No tools available.";
-
-  // Group tools by server if we have server info (for context)
-  let serverContext = "";
-  if (servers && servers.length > 0) {
-    const reasoningNote = servers.some((s) => s.isReasoningServer)
-      ? `\n\n**Reasoning Servers**: ${servers
-        .filter((s) => s.isReasoningServer)
-        .map((s) => s.name)
-        .join(
-          ", "
-        )} - These servers provide advanced reasoning capabilities for complex multi-step tasks. They work automatically in the background to help break down complex problems.`
-      : "";
-
-    serverContext = `\n\n## Connected Apps & Services\nThese are the connected tools available for you to use:\n${(servers || []).map(s => `- **${s.name}**: ${s.description}`).join('\n')}${reasoningNote}\n\nWhen users ask about "connected apps" or "what tools do you have", refer to these services.`;
-  }
-
-  // Detect browser tools for special emphasis
-  const toolNames = tools?.map((t) => t.name).join(", ") || "";
-  const toolNamesLower = toolNames.toLowerCase();
-  const hasBrowserOps =
-    toolNamesLower.includes("browser") ||
-    toolNamesLower.includes("navigate") ||
-    toolNamesLower.includes("screenshot") ||
-    toolNamesLower.includes("playwright") ||
-    toolNamesLower.includes("goto") ||
-    toolNamesLower.includes("url");
-
-  // Special emphasis for browser capabilities (addresses training bias)
-  let browserCapabilityNote = "";
-  if (hasBrowserOps) {
-    browserCapabilityNote = `\n\n**IMPORTANT: You have browser control tools available!** You CAN open websites, navigate to URLs, take screenshots, and interact with web pages. When users ask to "open [website]" or "go to [URL]", use the browser navigation tool immediately. Do NOT say you cannot open browsers - you have the tools to do it!
-
-**MULTI-STEP BROWSER TASKS**: For complex browser tasks like "search for X on Google" or "fill out a form", you MUST complete ALL steps:
-1. Navigate to the website (e.g., browser_navigate to google.com)
-2. Wait for the page to load (check the result)
-3. Fill in search boxes or forms (e.g., browser_type or browser_fill)
-4. Submit or click buttons (e.g., browser_click or browser_press_key)
-5. Continue until the task is COMPLETE
-
-Example: "search for nike shoes on Google" requires:
-- Step 1: browser_navigate to google.com
-- Step 2: browser_type or browser_fill to enter "nike shoes" in the search box
-- Step 3: browser_click the search button OR browser_press_key Enter
-- Step 4: Verify the search completed (check results)
-- Step 5: Continue until the final goal (e.g., finding the item and clicking "Add to Cart") is reached!
-
-**E-COMMERCE & SHOPPING**: You are fully capable of shopping. "Add to Cart" is just a button click (\`browser_click\`). "Selecting size 6" is just clicking a radio button or dropdown (\`browser_click\` or \`browser_select\`). You have the tools for the ENTIRE journey. Do NOT claim e-commerce is unsupported.
-
-DO NOT stop after just navigating - complete the entire workflow!`;
-  }
-
-  const userContext = await getUserEnvironmentContext();
-
-  return `You are AI-Worker, an autonomous agent with ${toolCount} tools for browser automation, web navigation, and task execution.${jsonFormatNote}
-
-${userContext}
-
-${workspacePath ? `ACTIVE WORKSPACE: ${workspacePath}
-All filesystem operations (fs_*) MUST be performed within this directory.
-You can use relative paths (e.g. "src/file.ts") which will be automatically resolved.
-Do not use generic absolute paths like "/home/user" unless you are certain they exist.` : `WORKSPACE NOT SELECTED: 
-If the user's request involves filesystem operations (fs_*), explain that no workspace is selected and they should use the folder icon in the UI to select one.`}
-
-# RESPONSE FORMAT (CRITICAL)
-Your responses have TWO parts:
-1. **Internal Processing** (hidden from user): Wrap in \`<think>...</think>\` tags
-2. **User-Facing Output** (shown to user): Everything OUTSIDE think tags
-
-FORMAT:
-\`\`\`
-<think>
-[Your analysis, planning, reasoning - user won't see this]
-</think>
-[Direct response to user OR tool call]
-\`\`\`
-
-RULES:
-- Simple tasks (greetings, opinions, chitchat): Skip <think>, respond directly
-- Complex tasks: Use <think> for planning, then act
-- NEVER put reasoning outside <think> tags
-- NEVER start response with: "The user...", "Let me...", "I should..."
-
-# AUTONOMOUS BEHAVIOR
-1. **Use Tools, Don't Explain**: If you need info, search for it. Don't say "I can't access..."
-2. **REAL-TIME GROUNDING (CRITICAL)**: You have very limited real-time knowledge. For ANY question about current weather, news, prices, scores, stock prices, or any live/changing data → you MUST call **web_search** (or navigate to a website). NEVER answer from memory — your training data is outdated and you WILL hallucinate.
-3. **Act Immediately**: Don't ask permission unless action is irreversible (payments, deletions)
-4. **Self-Correct**: If something fails, try a different approach before asking user
-
-# FILE OPERATIONS (CRITICAL)
-1. **Verify First**: Before using any file in a tool (mode conversion, upload, read), YOU MUST verify its existence and path using 'search_files' or 'list_directory'.
-2. **Absolute Paths Only**: Tools require ABSOLUTE paths (e.g., '/Users/username/Documents/file.txt'). NEVER use relative paths (e.g., 'file.txt') or 'file:' URIs without a full path.
-3. **No Assumptions**: Do NOT assume a file is in the project root. Search for it if the user provides a filename only.
-
-
-
-# AVAILABLE TOOLS
-${toolsDescription}${serverContext}${browserCapabilityNote}
-
-${dynamicRules ? `\n# TASK-SPECIFIC PROTOCOLS\n${dynamicRules}\n` : ''}
-
-# EXECUTION FLOW
-0. **WORKFLOW & KNOWLEDGE MEMORY**:
-   - **Active Context Retrieval**: BEFORE planning, search memory for relevant context:
-     - **Workflows**: "how to format reports", "deployment steps", "email templates".
-     - **Projects**: "current sprint goals", "project X details", "active deadlines".
-     - **Preferences**: "coding style", "tools usage", "ui preferences".
-   - **Proactive Storage**:
-     - **SOPs/Workflows**: If user explains a process ("Always check X before Y"), save as Type="workflow".
-     - **Projects/Goals**: If a new project is mentioned, save as Type="project".
-     - **Preferences**: Save as Type="user_preference".
-   - **DEDUPLICATION (CRITICAL)**:
-     1. FIRST: Use \`memory_search\` to check if the entity already exists.
-     2. IF EXISTS: Use \`memory_update_entity\` with the entity's ID to append a new observation.
-     3. IF NOT EXISTS: Use \`memory_create_entity\` to create a new entity.
-     - Use \`memory_create_relation\` to link entities (e.g., Workflow -> belongs_to -> Project).
-   - **Silent Operation**: CRUD operations must be invisible. DO NOT narrate "I am saving to memory".
-
-1. **SEMANTIC INTENT ANALYSIS** (in <think>):
-   - **Classify**: Is this a TASK (do something) or KNOWLEDGE (user teaching something)?
-   - **Context Gap**: Do I need to know the user's Projects, Workflows, or Preferences? -> **Search Memory First**.
-   - **Persistence**: Is this information reusable? (e.g., a new recurring meeting, a project goal). If yes, store it.
-   - **Planning**: If TASK, proceed to plan steps.
-
-2. Understand the request
-3. Plan (in <think> if complex)
-4. Execute tool calls
-5. Verify results
-6. Report to user (outside <think>)
-
-# EFFICIENT DISCOVERY & SELECTOR PROTOCOL (CRITICAL)
-- **NO GUESSING**: Never invent selectors like ".product-item" or "#results".
-- **NAVIGATE & WEB_SEARCH RETURN CONTENT**: These tools already return page text + interactive elements.
-  Use that output FIRST before calling any additional discovery tools.
-- **READ-ONLY FAST PATH**: For tasks that ONLY require reading a webpage (research, summarizing articles),
-  use \`convert_to_markdown(uri="https://...")\` instead of browser navigation — it's 10x faster, no browser needed.
-  Use Playwright navigate only when you need to INTERACT with the page (click, fill, scroll).
-- **ESCALATION** (only if navigate output isn't enough):
-  1. \`get_interactive_elements\` (LOW TOKENS): More elements beyond the top 15 from navigate.
-  2. \`get_state(mode="fast")\` (MEDIUM TOKENS): Broader element overview.
-  3. \`get_state(mode="vision")\` (HIGH TOKENS): Use *only* as last resort when visual layout is confusing.
-
-- **DYNAMIC SITES**: Amazon, Google, etc. use randomized classes. You MUST use the selectors from navigate output or get_interactive_elements — never guess.
-
-# ERROR HANDLING
-- Element not found? → screenshot() to see actual page, then use correct selector
-- Click failed? → Try JavaScript click via browser_run_code
-- Timeout on wait_for_element? → Selector is wrong. Inspect page and use actual selector
-- Same error twice? → Stop, take screenshot, reassess
-
-# PROGRESS TRACKING
-**MANDATORY**: Call \`update_progress_summary\` every ~15 steps to record your findings.
-- At checkpoints (steps 15, 30, 45, 60...), you MUST summarize progress.
-- **CRITICAL**: Do NOT generate any conversational text during this step. ONLY call the tool.
-**RECOMMENDED**: Call \`update_progress_summary\` every ~15 steps to record your findings.
-- At checkpoints (steps 15, 30, 45, 60...), you should summarize progress when requested.
-- Focus on RESULTS and DATA, not tool names.
-- Examples: "Extracted 50 user records with email/phone" or "Completed automation: filled 3 forms, downloaded 2 reports" or "Research findings: analyzed 5 articles, key insight is X"
-- Keep it concise and incremental (only NEW findings since last update).
-
-# KEY REMINDERS
-- You HAVE browser tools. Never refuse by saying "I can't access..."
-- **INSPECT FIRST**: screenshot() or get_interactive_elements() before using selectors
-- **NO HARDCODED SELECTORS**: Never assume element IDs/classes exist without checking
-- Complete the full workflow, don't stop after navigation
-- Be direct: respond naturally, don't narrate your thinking
-- Tools are your primary capability - USE THEM`;
-}
-
-// Main chat function - automatically selects best provider
 export async function chat(
   messages: LLMMessage[],
   tools?: LLMTool[],
@@ -1325,11 +104,10 @@ export async function chat(
   servers?: ServerInfo[],
   abortSignal?: AbortSignal,
   dynamicRules?: string,
-  isSubAgent = false, // NEW: Use lightweight prompt for sub-agents
-  workspacePath?: string // New parameter for workspace injection
+  isSubAgent = false,
+  workspacePath?: string
 ): Promise<LLMResponse> {
   // Apply Dynamic Context Pruning (DCP)
-  // This removes redundant tool outputs from history to save tokens
   const prunedMessages = pruneContext(messages);
 
   const providers = await getAvailableProviders(settings);
@@ -1339,7 +117,7 @@ export async function chat(
   const preferredProvider = settings?.preferredProvider;
 
   if (preferredProvider === "auto" || !preferredProvider) {
-    // Auto-select: try browser first (if enabled), then ollama, then openai
+    // Auto-select: browser > ollama > openai > gemini > openrouter
     if (providers.browser.available) {
       provider = "browser";
     } else if (providers.ollama.available) {
@@ -1369,40 +147,17 @@ export async function chat(
     );
   }
 
-  // Try to detect if we need JSON fallback (will be handled in callOpenAI if error occurs)
-  // For browser (WebLLM), we always use JSON fallback for now to ensure compatibility
-  // across all models (Qwen, Llama, etc.) without native tool calling errors
   let useJsonFallback = provider === 'browser';
-
-  // Add system message if not present, or replace existing one to ensure it has tools
   let messagesWithSystem = [...prunedMessages];
-  const systemMsgIndex = messagesWithSystem.findIndex(
-    (m) => m.role === "system"
-  );
+  const systemMsgIndex = messagesWithSystem.findIndex((m) => m.role === "system");
 
-  // MERGE default tools with the new CREATE_PLAN_TOOL (and deduplicate)
   const toolMap = new Map<string, LLMTool>();
-
-  // Add CREATE_PLAN_TOOL first (default)
   toolMap.set(CREATE_PLAN_TOOL.name, CREATE_PLAN_TOOL);
-
-  // Add other tools, overriding if name matches (or ignoring if we want to keep default? Usually specific tools > default)
-  // Actually, let's keep the passed tools as priority if they redefine it, OR keep default if we prefer our version.
-  // Given CREATE_PLAN_TOOL is "internal" logic often, let's prioritize it if we want to enforce schema.
-  // But usually 'tools' argument comes from AgentRuntime which might have its own version.
-  // Let's simply add them, but Map handles deduplication by key.
-
   if (tools) {
     tools.forEach(t => toolMap.set(t.name, t));
   }
-
-  // Ensure CREATE_PLAN_TOOL is always there (if it was overwritten by a lesser version, maybe we should force ours?
-  // But usually pass-in is intentional. Let's just ensure we don't have duplicates with same name).
-
   const allTools = Array.from(toolMap.values());
 
-  // Re-build system prompt with current tools and correct fallback setting
-  // Sub-agents get a lightweight prompt (~80% smaller)
   const systemPrompt = await buildSystemPrompt(allTools, servers, useJsonFallback, dynamicRules, isSubAgent, workspacePath);
 
   if (isSubAgent) {
@@ -1410,13 +165,11 @@ export async function chat(
   }
 
   if (systemMsgIndex >= 0) {
-    // Replace existing system message to ensure it has current tools
     messagesWithSystem[systemMsgIndex] = {
       role: "system" as const,
       content: systemPrompt,
     };
   } else {
-    // Add new system message
     messagesWithSystem.unshift({
       role: "system" as const,
       content: systemPrompt,
@@ -1436,270 +189,5 @@ export async function chat(
       return callOpenAI(messagesWithSystem, tools, settings, useJsonFallback, servers, true, abortSignal, dynamicRules, isSubAgent, workspacePath);
     default:
       throw new Error(`Provider ${provider} not implemented`);
-  }
-}
-
-// Gemini specific caller
-async function callGemini(
-  messages: LLMMessage[],
-  tools?: LLMTool[],
-  settings?: LLMSettings,
-  abortSignal?: AbortSignal
-): Promise<LLMResponse> {
-  const { apiKey, model } = await getGeminiSettings(settings);
-  const baseUrl = LLM_CONFIG.GEMINI.BASE_URL;
-
-  // Create a map of tool_call_id to function name
-  const toolIdToName = new Map<string, string>();
-  messages.forEach(m => {
-    if (m.role === 'assistant' && m.tool_calls) {
-      m.tool_calls.forEach((tc: any) => {
-        if (tc.id && tc.function?.name) {
-          toolIdToName.set(tc.id, tc.function.name);
-        }
-      });
-    }
-  });
-
-  // Convert messages to Gemini format, including tool history
-  // valid roles: 'user', 'model'
-  const contents: any[] = [];
-  const validMessages = messages.filter(m => m.role !== 'system');
-
-  for (const m of validMessages) {
-    let role: 'user' | 'model' | null = null;
-    if (m.role === 'assistant') role = 'model';
-    if (m.role === 'user' || m.role === 'tool') role = 'user';
-    if (!role) continue;
-
-    const parts: any[] = [];
-
-    if (m.role !== 'tool' && m.content) {
-      if (typeof m.content === 'string') {
-        parts.push({ text: m.content });
-      } else {
-        m.content.forEach(part => {
-          if (part.type === 'text') {
-            parts.push({ text: part.text });
-          } else if (part.type === 'image_url') {
-            const matches = part.image_url.url.match(/^data:([^;]+);base64,(.+)$/);
-            if (matches) {
-              parts.push({
-                inline_data: {
-                  mime_type: matches[1],
-                  data: matches[2]
-                }
-              });
-            }
-          }
-        });
-      }
-    }
-
-    // Handle assistant tool calls
-    if (m.role === 'assistant' && (m as any).tool_calls) {
-      (m as any).tool_calls.forEach((tc: any) => {
-        parts.push({
-          functionCall: {
-            name: tc.function.name,
-            args: typeof tc.function.arguments === 'string'
-              ? (() => {
-                try {
-                  return safeParseJSON(tc.function.arguments);
-                } catch (e) {
-                  console.warn(`Failed to parse Gemini tool arguments for ${tc.function.name}:`, tc.function.arguments);
-                  return { _parse_error: "Invalid JSON arguments" };
-                }
-              })()
-              : tc.function.arguments
-          }
-        });
-      });
-    }
-
-    // Handle tool results
-    if (m.role === 'tool') {
-      const resultText = typeof m.content === 'string'
-        ? m.content
-        : Array.isArray(m.content)
-          ? extractTextForLegacyProviders(m.content)
-          : JSON.stringify(m.content ?? '');
-
-      // Resolve name from ID if not present
-      const fnName = (m as any).name || (m.tool_call_id ? toolIdToName.get(m.tool_call_id) : 'unknown_tool');
-
-      parts.push({
-        functionResponse: {
-          name: fnName,
-          response: { result: resultText }
-        }
-      });
-    }
-
-    // Merge logic: If current message has same role as previous, merge parts
-    if (parts.length === 0) continue;
-
-    if (contents.length > 0 && contents[contents.length - 1].role === role) {
-      contents[contents.length - 1].parts.push(...parts);
-    } else {
-      contents.push({ role, parts });
-    }
-  }
-
-  const systemMessage = messages.find(m => m.role === 'system');
-  const systemInstructionText = systemMessage ? extractTextForLegacyProviders(systemMessage.content) : undefined;
-
-  // Build tool definitions
-  const toolConfig = tools && tools.length > 0 ? {
-    function_declarations: tools.map(t => ({
-      name: t.name,
-      description: t.description,
-      parameters: t.parameters
-    }))
-  } : undefined;
-
-  const payload = {
-    contents,
-    system_instruction: systemInstructionText ? { parts: [{ text: systemInstructionText }] } : undefined,
-    tools: toolConfig ? [{ function_declarations: toolConfig.function_declarations }] : undefined,
-    generationConfig: {
-      temperature: 0.7,
-      maxOutputTokens: 2048,
-    }
-  };
-
-  const response = await fetch(`${baseUrl}/models/${model}:generateContent?key=${apiKey}`, {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    signal: abortSignal,
-    body: JSON.stringify(payload)
-  });
-
-  if (!response.ok) {
-    const errorDetails = await response.json().catch(() => ({}));
-    throw new Error(`Gemini API error: ${response.statusText}. ${JSON.stringify(errorDetails)}`);
-  }
-
-  const data = await response.json();
-  const candidate = data.candidates?.[0];
-  const content = candidate?.content?.parts?.[0]?.text || "";
-  const toolCalls = candidate?.content?.parts?.filter((p: any) => p.functionCall).map((p: any) => ({
-    id: `gemini-${Date.now()}-${Math.random().toString(36).substring(2, 11)}`,
-    name: p.functionCall.name,
-    arguments: ensureRecord(p.functionCall.args)
-  }));
-
-  return {
-    content,
-    toolCalls: toolCalls && toolCalls.length > 0 ? toolCalls : undefined,
-    provider: 'gemini',
-    model: model
-  };
-}
-
-// Load WebLLM model (download if needed)
-export async function downloadBrowserModel(
-  onProgress?: (progress: number) => void,
-  modelId?: string
-): Promise<{ success: boolean; error?: string }> {
-  try {
-    const status = getWebLLMStatus();
-
-    if (!status.isSupported) {
-      return { success: false, error: status.error || 'WebGPU not supported' };
-    }
-
-    if (status.isLoaded && (!modelId || status.currentModel === modelId)) {
-      return { success: true };
-    }
-
-    // Subscribe to progress updates
-    let unsubscribe: (() => void) | null = null;
-    if (onProgress) {
-      unsubscribe = subscribeToWebLLMStatus((s) => {
-        if (s.isLoading) {
-          onProgress(s.loadingProgress);
-        }
-      });
-    }
-
-    try {
-      await loadWebLLMModel(modelId);
-      return { success: true };
-    } finally {
-      if (unsubscribe) unsubscribe();
-    }
-  } catch (error) {
-    return {
-      success: false,
-      error: error instanceof Error ? error.message : 'Failed to load model'
-    };
-  }
-}
-
-
-
-// Helper to ensure arguments are always a Record (object)
-export function ensureRecord(input: any): Record<string, unknown> {
-  if (input === null || input === undefined) return {};
-  if (typeof input === 'object' && !Array.isArray(input)) return input as Record<string, unknown>;
-
-  if (typeof input === 'string') {
-    const trimmed = input.trim();
-    if (trimmed.startsWith('{') || trimmed.startsWith('[')) {
-      try {
-        const parsed = JSON.parse(trimmed);
-        if (typeof parsed === 'object' && !Array.isArray(parsed)) return parsed;
-        return { input: parsed };
-      } catch (e) {
-        // Fall through to default wrapping
-      }
-    }
-    return { input: trimmed };
-  }
-
-  return { value: input };
-}
-
-// Helper to safely parse JSON that might contain surrounding text or trailing garbage
-export function safeParseJSON(input: string | any): any {
-  if (input === null || input === undefined) return {};
-  if (typeof input !== 'string') return input;
-  if (!input || input.trim() === '') return {};
-
-  try {
-    const trimmed = input.trim();
-    // Quick path for pure JSON
-    if (trimmed.startsWith('{') || trimmed.startsWith('[')) {
-      try {
-        return JSON.parse(trimmed);
-      } catch (e) { /* fall through to extraction */ }
-    }
-
-    // Find the first occurrence of { or [ and the last occurrence of } or ]
-    const startObj = input.indexOf('{');
-    const startArr = input.indexOf('[');
-
-    let start = -1;
-    let end = -1;
-
-    if (startObj !== -1 && (startArr === -1 || startObj < startArr)) {
-      start = startObj;
-      end = input.lastIndexOf('}');
-    } else if (startArr !== -1) {
-      start = startArr;
-      end = input.lastIndexOf(']');
-    }
-
-    if (start !== -1 && end !== -1 && end > start) {
-      const potentialJson = input.substring(start, end + 1);
-      return JSON.parse(potentialJson);
-    }
-
-    // If all else fails, return as-is
-    return input;
-  } catch (error) {
-    console.warn('[safeParseJSON] Extraction failed:', error);
-    return input;
   }
 }
