+++ src/renderer/src/lib/llm/types.ts
@@ -1,0 +1,16 @@
+import { LLMMessage, LLMSettings, LLMTool, LLMProvider, LLMResponse, ServerInfo, LLMContentPart } from "../types";
+
+export interface ProviderStatus {
+  available: boolean;
+  model?: string;
+  error?: string;
+  models?: string[]; // Available models list
+  modelsEndpointAvailable?: boolean; // Whether /models endpoint exists
+  // WebLLM specific
+  isWebGPUSupported?: boolean;
+  isLoaded?: boolean;
+  isLoading?: boolean;
+  loadingProgress?: number;
+  loadingStage?: string;
+  downloadedModels?: string[]; // WebLLM cached models
+}
